homepage dashboards now change when filtering
