+++ trakky-client/src/app/expenses/components/edit-cell.tsx
@@ -1,21 +1,20 @@
 import { PopupDialog } from "@/app/expenses/components/popup-dialog.tsx";
 import { Button } from "@/components/ui/button.tsx";
 import { PenBoxIcon } from "lucide-react";
-import { PaymentForm } from "@/app/expenses/components/payment-form.tsx";
-import { Payment } from "@/infrastructure/payment.tsx";
+import React from "react";
 
 export function EditCell({
-  payment,
-  onPaymentEdited,
+  id,
+  children
 }: {
-  payment: Payment;
-  onPaymentEdited: () => void;
+  id: any;
+  children: React.ReactNode;
 }) {
   return (
     <PopupDialog
       trigger={
         <Button
-          key={payment.id}
+          key={id}
           variant="outline"
           className="bg-transparent hover:bg-transparent p-0 mx-1 my-0 h-5 border-none hover:text-green-500"
         >
@@ -27,11 +26,7 @@ export function EditCell({
         </Button>
       }
     >
-      <PaymentForm
-        editValues={payment}
-        refresh={onPaymentEdited}
-        title={"Edit Transaction"}
-      ></PaymentForm>
+      {children}
     </PopupDialog>
   );
 }
