+++ trakky-client/src/app/expenses/components/table.tsx
@@ -1,19 +1,10 @@
 "use client";
 
-import { useEffect, useMemo, useState } from "react";
 import {
-  ColumnDef,
-  ColumnFiltersState,
   flexRender,
-  getCoreRowModel,
-  getFacetedMinMaxValues,
-  getFacetedRowModel,
-  getFacetedUniqueValues,
-  getFilteredRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  useReactTable,
 } from "@tanstack/react-table";
+
+import { Table as TableType } from "@tanstack/react-table";
 
 import {
   Table,
@@ -22,144 +13,45 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-import { Summary, Total } from "@/app/expenses/components/summary";
-import { DeletePayments, Payment } from "@/infrastructure/payment";
-import { fuzzyFilter } from "@/lib/filters";
+import { Payment } from "@/infrastructure/payment";
 import { cn } from "@/lib/utils.ts";
 import { Filter } from "@/app/expenses/components/column-filter.tsx";
 import {
   colSize,
-  ColumnDefinition,
 } from "@/app/expenses/components/columns.tsx";
 import { TableActionMenu } from "@/app/expenses/components/table-action-menu.tsx";
 import { EditCell } from "@/app/expenses/components/edit-cell.tsx";
 import { FadeUp } from "@/components/animations/fade.tsx";
+import { PaymentForm } from "@/app/expenses/components/payment-form.tsx";
 
-export interface DataTableProps {
-  data: Payment[] | null;
-  selectedYear: string | null;
-  refreshData: () => void;
+export interface ExpensesTableProps {
+  table: TableType<any>;
+  onDeleteConfirmed: () => Promise<void>;
+  onPaymentEdited: () => void;
+  onRefresh: () => Promise<void>;
 }
 
 export function ExpensesTable({
-  dataTableProps,
+  expensesTableProps,
   ...props
 }: {
-  dataTableProps: DataTableProps;
+  expensesTableProps: ExpensesTableProps;
 }) {
-  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
-  const [globalFilter, setGlobalFilter] = useState("");
-  const [filteredData, setFilteredData] = useState<Payment[]>([]);
-  const [columnVisibility, setColumnVisibility] = useState({});
-
-  useEffect(() => {
-    if (dataTableProps.selectedYear === "All" && dataTableProps.data)
-      setFilteredData(dataTableProps.data);
-    else if (dataTableProps.data && dataTableProps.selectedYear !== null) {
-      setFilteredData(
-        dataTableProps.data.filter(
-          (payment) =>
-            new Date(payment.date).getFullYear() ===
-            parseInt(dataTableProps.selectedYear ?? ""),
-        ),
-      );
-    }
-  }, [dataTableProps.data, dataTableProps.selectedYear]);
-
-  const totalsPerYear =
-    dataTableProps.data === null
-      ? []
-      : dataTableProps.data.reduce((acc: Total[], payment) => {
-        const year = new Date(payment.date).getFullYear();
-        const existing = acc.find((t) => t.date === year);
-        if (existing) {
-          existing.amount += payment.amount;
-        } else {
-          acc.push({ amount: payment.amount, date: year });
-        }
-        return acc;
-      }, []);
-
-  const columns = useMemo<ColumnDef<Payment, number | string>[]>(
-    () => ColumnDefinition,
-    [],
-  );
-
-  const table = useReactTable({
-    data: filteredData,
-    columns,
-    filterFns: {
-      fuzzy: fuzzyFilter,
-    },
-    state: {
-      columnFilters,
-      globalFilter,
-      columnVisibility,
-    },
-    onColumnVisibilityChange: setColumnVisibility,
-    onColumnFiltersChange: setColumnFilters,
-    onGlobalFilterChange: setGlobalFilter,
-    globalFilterFn: fuzzyFilter,
-    getCoreRowModel: getCoreRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    getFacetedRowModel: getFacetedRowModel(),
-    getFacetedUniqueValues: getFacetedUniqueValues(),
-    getFacetedMinMaxValues: getFacetedMinMaxValues(),
-    debugTable: true,
-    debugHeaders: true,
-    debugColumns: false,
-  });
-
-  async function onDeleteConfirmed() {
-    console.log("Delete clicked!");
-    const ids = table
-      .getSelectedRowModel()
-      .rows.map((row: any) => row.original.id) as number[];
-
-    console.log(ids);
-
-    const deleted = await DeletePayments(ids);
-
-    if (deleted) {
-      table.resetRowSelection();
-      dataTableProps.refreshData();
-      alert("Transactions deleted!");
-    } else {
-      alert("Error! Could not delete transactions");
-    }
-  }
-
-  function onPaymentEdited() {
-    table.resetRowSelection();
-    onRefresh().then(() => { });
-  }
-
-  async function onRefresh() {
-    table.resetColumnFilters();
-    dataTableProps.refreshData();
-  }
 
   return (
     <div {...props}>
       {
         <>
-          <Summary
-            table={table}
-            totalsPerYear={totalsPerYear}
-            selectedYear={dataTableProps.selectedYear ?? ""}
-          />
           <FadeUp>
             <TableActionMenu
-              table={table}
-              onDeleteConfirmed={onDeleteConfirmed}
-              onRefresh={onRefresh}
+              table={expensesTableProps.table}
+              onDeleteConfirmed={expensesTableProps.onDeleteConfirmed}
+              onRefresh={expensesTableProps.onRefresh}
             />
             <div>
               <Table className="bg-slate-950 border border-slate-800 overflow-x-scroll">
                 <TableHeader className="hover:bg-transparent">
-                  {table.getHeaderGroups().map((headerGroup) => (
+                  {expensesTableProps.table.getHeaderGroups().map((headerGroup) => (
                     <TableRow
                       key={headerGroup.id}
                       className="border border-slate-800"
@@ -206,7 +98,7 @@ export function ExpensesTable({
                                   <div className="m-0 p-0">
                                     <Filter
                                       column={header.column}
-                                      table={table}
+                                      table={expensesTableProps.table}
                                     />
                                   </div>
                                 ) : null}
@@ -219,7 +111,7 @@ export function ExpensesTable({
                   ))}
                 </TableHeader>
                 <TableBody>
-                  {table.getRowModel().rows.map((row) => {
+                  {expensesTableProps.table.getRowModel().rows.map((row) => {
                     return (
                       <TableRow
                         key={row.id}
@@ -249,10 +141,13 @@ export function ExpensesTable({
                           );
                         })}
                         <td>
-                          <EditCell
-                            onPaymentEdited={onPaymentEdited}
-                            payment={row.original as Payment}
-                          />
+                          <EditCell id={(row.original as Payment).id}>
+                            <PaymentForm
+                              editValues={row.original as Payment}
+                              refresh={expensesTableProps.onPaymentEdited}
+                              title={"Edit Transaction"}
+                            ></PaymentForm>
+                          </EditCell>
                         </td>
                       </TableRow>
                     );
