+++ trakky-client/src/app/expenses/page.tsx
@@ -3,6 +3,9 @@ import { ExpensesTable } from "./components/table";
 import { Text } from "@/components/ui/text.tsx";
 import { usePaymentData } from "@/lib/hooks.ts";
 import { YearSelection } from "@/components/ui/data-selector.tsx";
+import { useTable } from "@/lib/table-hooks.ts";
+import { Summary } from "@/app/expenses/components/summary.tsx";
+
 
 export default function ExpensesPage() {
   const {
@@ -13,6 +16,18 @@ export default function ExpensesPage() {
     setSelectedYear,
   } = usePaymentData();
 
+  const {
+    totalsPerYear,
+    table,
+    onDeleteConfirmed,
+    onPaymentEdited,
+    onRefresh,
+  } = useTable({
+    data: payments,
+    selectedYear,
+    refreshData,
+  })
+
   return (
     <PageContainer>
       <Text title={"Expenses"} />
@@ -21,11 +36,17 @@ export default function ExpensesPage() {
         selectedYear={selectedYear}
         onYearChange={setSelectedYear}
       />
+      <Summary
+        table={table}
+        totalsPerYear={totalsPerYear}
+        selectedYear={selectedYear ?? ""}
+      />
       <ExpensesTable
-        dataTableProps={{
-          data: payments,
-          refreshData: refreshData,
-          selectedYear: selectedYear,
+        expensesTableProps={{
+          table,
+          onDeleteConfirmed,
+          onPaymentEdited,
+          onRefresh,
         }}
       />
     </PageContainer>
