+++ trakky-client/src/components/ui/text.tsx
@@ -26,7 +26,7 @@ export function SubTitle({ title, ...props }: { title?: string }) {
             title={title}
             className="flex flex-col items-center w-full justify-center bg-transparent p-0 md:p-6"
           >
-            <p className="w-full p-5 pt-0 text-sm bg-transparent">{title}</p>
+            <p className="w-full px-5 pt-0 text-sm bg-transparent text-slate-500">{title}</p>
           </CardTitle>
         </CardHeader>
       </Card>
