+++ trakky-client/src/constants.ts
@@ -1,3 +1,3 @@
 export const serverUrl = "http://trakky-api.localdomain";
 
-export const demoMode = true;
+export const demoMode = false;
